rag_simple/requirements.txt
# RAG — работает без OpenAI кредитов
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-groq>=0.2.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
pypdf>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.37.0
